renaming from profiling.rs to metrics.rs, and the example metrics server/client
