+++ lib/runtime/examples/system_metrics/Cargo.toml
@@ -14,7 +14,7 @@
 # limitations under the License.
 
 [package]
-name = "system_stats_endpoint"
+name = "system_metrics"
 version.workspace = true
 edition.workspace = true
 authors.workspace = true
