+++ lib/runtime/examples/system_metrics/src/lib.rs
