+++ lib/runtime/src/component.rs
@@ -30,7 +30,7 @@
 //! TODO: Top-level Overview of Endpoints/Functions
 
 use crate::{
-    discovery::Lease, profiling::MetricsRegistry, service::ServiceSet, transports::etcd::EtcdPath,
+    discovery::Lease, metrics::MetricsRegistry, service::ServiceSet, transports::etcd::EtcdPath,
 };
 
 use super::{
