+++ lib/runtime/src/lib.rs
@@ -38,9 +38,9 @@ pub mod discovery;
 pub mod engine;
 pub mod http_server;
 pub mod logging;
+pub mod metrics;
 pub mod pipeline;
 pub mod prelude;
-pub mod profiling;
 pub mod protocols;
 pub mod runnable;
 pub mod runtime;
